feat: Finish 02-flex-header exercise
This exercise was slightly challenging as I had to
reset the margins and padding of the entire document
(and I totally forgot about it after watching
Slaying the Dragon's CSS Box Model video).

I also put in "box-sizing: border box;" in the
universal selector to create a habit of writing
them in every CSS file.

Additionally, I found that using flex is *kinda*
intuitive in styling elements in the webpage.  I
need more practice though! :)